// src/app/google/types/google/calendar.ts
export type date = string;

/**
 * วันเวลาและวันที่ รูปแบบรวมกัน (ตามมาตรฐาน RFC 3339)
 * ต้องระบุ Time zone offset เสมอ ยกเว้นจะใช้ฟิลด์ timeZone แยกต่างหาก
 */
export type datetime = string;

export interface NestedDateTime {
  /** The time, as a combined date-time value (formatted according to RFC 3339). */
  readonly dateTime: datetime;
  /** The time zone in which the time is specified. (e.g. "Europe/Zurich"). */
  readonly timeZone: string;
}

export interface NestedDateOnly {
  /** The date, in the format "yyyy-mm-dd", as defined by RFC 3339. */
  readonly date: date;
  /** The time zone in which the time is specified. */
  readonly timeZone: string;
}

/**
 * ตัวแทนเวลาเริ่มหรือสิ้นสุดของกิจกรรม โดยเลือกได้ว่าจะเป็นแค่วันที่ (NestedDateOnly) หรือเต็มรูปแบบ (NestedDateTime)
 */
export type NestedDate = NestedDateTime | NestedDateOnly;
// ...
export interface NestedDateOnlyRequired extends DeepPartial<NestedDateOnly> {
  /** The date, in the format "yyyy-mm-dd", as defined by RFC 3339. */
  readonly date: NestedDateOnly['date'];
}

/**
 * Required nested date-time object for events.insert.
 */
export interface NestedDateTimeRequired extends DeepPartial<NestedDateTime> {
  /** The time, as a combined date-time value (formatted according to RFC 3339). */
  readonly dateTime: NestedDateTime['dateTime'];
}

/**
 * Union of required start/end time formats for insertion.
 */
export type NestedDateRequired = NestedDateOnlyRequired | NestedDateTimeRequired;

/**
 * Request body for inserting an event.
 */
export interface EventResourceInsertBody extends DeepPartial<EventResource> {
  /** The (inclusive) start time of the event. */
  readonly start: NestedDateRequired;
  /** The (exclusive) end time of the event. */
  readonly end: NestedDateRequired;
}